JPA替换

+++ src/main/java/com/site/springboot/core/controller/admin/CategoryController.java
@@ -92,7 +92,7 @@ public class CategoryController {
      */
     @RequestMapping(value = "/categories/delete", method = RequestMethod.POST)
     @ResponseBody
-    public Result delete(@RequestBody Integer[] ids) {
+    public Result delete(@RequestBody Long[] ids) {
         if (ids.length < 1) {
             return ResultGenerator.genFailResult("参数异常！");
         }
